r10718/sofa : push 516650b..2a858c1
* commit 2a858c15f3f494f7010ec73246fe2aae91e26ada
  Author: Marc Legendre <marc.legendre@inria.fr>
  Date:   Tue Feb 25 20:25:51 2014 +0100
  
      Fix minor warnings from GCC
      
      This commit also silences the 'unused local typedef' warnings (introduced in
      GCC 4.8) from miniboost with __attribute__((unused)).
  
   extlibs/miniBoost/boost/bind/arg.hpp               |    4 ++++
   extlibs/miniBoost/boost/concept/detail/general.hpp |    9 +++++++++
   .../miniBoost/boost/graph/breadth_first_search.hpp |    4 ++++
   .../miniBoost/boost/graph/connected_components.hpp |    4 ++++
   .../boost/graph/cuthill_mckee_ordering.hpp         |   12 ++++++++++++
   extlibs/miniBoost/boost/graph/graph_utility.hpp    |   10 ++++++++++
   .../miniBoost/boost/tuple/detail/tuple_basic.hpp   |    4 ++++
   .../miniBoost/boost/unordered/detail/unique.hpp    |    4 ++++
   .../sofa/component/collision/RayTraceDetection.cpp |    6 ------
   .../component/constraintset/LMConstraintSolver.cpp |    1 -
   .../mapping/CenterOfMassMulti2Mapping.inl          |    2 --
   .../BezierTetrahedronSetTopologyContainer.cpp      |   10 +++++-----
   12 files changed, 56 insertions(+), 14 deletions(-)



Former-commit-id: 49effaa593f78e93ac9417fef1ccd01bdfd2a12b

+++ modules/sofa/component/topology/BezierTetrahedronSetTopologyContainer.cpp
@@ -276,7 +276,7 @@ void BezierTetrahedronSetTopologyContainer::getGlobalIndexArrayOfBezierPointsInT
 			offset=getNumberOfTetrahedralPoints()+eit[i]*(degree-1);
 			// check the order of the edge to be consistent with the tetrahedron
 			if (e[0]==tet[edgesInTetrahedronArray[i][0]]) {
-				for (j=0;j<(degree-1);++j) {
+				for (j=0;j<(size_t)(degree-1);++j) {
 					indexArray.push_back(offset+j);
 				}
 			} else {
@@ -304,7 +304,7 @@ void BezierTetrahedronSetTopologyContainer::getGlobalIndexArrayOfBezierPointsInT
 				indexTriangle[(k+2)%3]=1;
 				indexTriangle[(k+1)%3]=2;
 			}
-			for (j=1;j<(degree-1);++j) {
+			for (j=1;j<(size_t)(degree-1);++j) {
 				for (k=1;k<(degree-j);++k) {
 					TetrahedronBezierIndex bti(0,0,0,0);
 					bti[trianglesInTetrahedronArray[i][indexTriangle[0]]]=j;
@@ -323,7 +323,7 @@ void BezierTetrahedronSetTopologyContainer::getGlobalIndexArrayOfBezierPointsInT
 		size_t pointsPerTetrahedron=(degree-1)*(degree-2)*(degree-3)/6;
 		offset=getNumberOfTetrahedralPoints()+getNumberOfEdges()*(degree-1)+getNumberOfTriangles()*(degree-1)*(degree-2)/2+tetrahedronIndex*pointsPerTetrahedron;
 		size_t rank=0;
-		for (i=0;i<(degree-3);++i) {
+		for (i=0;i<(size_t)(degree-3);++i) {
 			for (j=0;j<(degree-i-3);++j) {
 				for (k=0;k<(degree-j-i-3);++k) {
 					indexArray.push_back(offset+rank);
@@ -408,11 +408,11 @@ bool BezierTetrahedronSetTopologyContainer::checkBezierPointTopology()
 				assert(elementIndex==getTetrahedron(nTetras)[elem]);
 				assert(elementOffset==0);
 			}
-			else if (elem<(4+6*(degree-1))){
+			else if (elem<(size_t)(4+6*(degree-1))){
 				assert(location==EDGE);
 				assert(elementIndex==getEdgesInTetrahedron(nTetras)[(elem-4)/(degree-1)]);
 			}
-			else if (elem<(4+6*(degree-1)+2*(degree-1)*(degree-2))){
+			else if (elem<(size_t)(4+6*(degree-1)+2*(degree-1)*(degree-2))){
                 assert(location==TRIANGLE);
 #ifndef NDEBUG
                 size_t nbPointPerEdge=(degree-1)*(degree-2)/2;
